feat(web): add getInfosBySlug and getSeriesChapters
