+++ web/app/features/asura/components/Modal/Delete.test.ts
@@ -2,7 +2,7 @@
 // @vitest-environment nuxt
 
 import type { VueWrapper } from '@vue/test-utils'
-import type { AsuraSearchItem } from '../../composables/asura.api'
+import type { AsuraSearchItem } from '../../types'
 import { mountSuspended } from '@nuxt/test-utils/runtime'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { defineComponent, h, ref } from 'vue'
