+++ web/app/features/asura/components/Modal/Delete.vue
@@ -1,5 +1,7 @@
 <!-- SPDX-License-Identifier: AGPL-3.0-or-later -->
 <script setup lang="ts">
+import type { AsuraSearchItem } from '../../types'
+
 const show = defineModel<boolean>({ required: true })
 const comic = defineModel<AsuraSearchItem | undefined>('comic', { required: true })
