+++ web/app/features/asura/composables/asura-search.composable.test.ts
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // @vitest-environment nuxt
 
-import type { AsuraSearchItem } from './asura.api'
+import type { AsuraSearchItem } from '../types'
 import { mockNuxtImport } from '@nuxt/test-utils/runtime'
 import { createPinia, setActivePinia } from 'pinia'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
