+++ web/app/pages/browse/sources/asura/index.test.ts
@@ -2,7 +2,7 @@
 // @vitest-environment nuxt
 
 import type { VueWrapper } from '@vue/test-utils'
-import type { AsuraSearchItem } from '~/features/asura/composables/asura.api'
+import type { AsuraSearchItem } from '~/features/asura/types'
 import { mountSuspended } from '@nuxt/test-utils/runtime'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { defineComponent, h } from 'vue'
